new class DictFilterNode created due to duplication issue
